Remove duplication of python files
